+++ move/ts_scripts/src/config.ts
