+++ move/ts_scripts/src/delete_task.ts
@@ -1,19 +1,35 @@
-// npx ts-node ./ts/create_task.ts
+// npx ts-node ./src/delete_task.ts
 import { client } from "./utils";
 import { MODULE_ADDRESS } from "./config";
 import { getAccount } from "./account";
+import { getTaskTableHandle } from "./get_task_table";
+import { type Task } from "./types";
 
 async function main() {
   try {
+    // get tasks from chain
+    let tasks: Task[] = await getTaskTableHandle();
+
+    console.log(tasks);
+
+    // find task to delete (for this example, we delete the first task)
+    const taskToDelete = tasks[0]; // Change this to your desired task selection logic
+
+    if (!taskToDelete) {
+      console.log('No task to delete');
+      return;
+    }
+    
+    // delete task
     const account = await getAccount();
 
     // build transaction
     const transaction = await client.transaction.build.simple({
       sender: account.accountAddress,
       data: {
-        function: `${MODULE_ADDRESS}::todolist::create_task`,
-        functionArguments: ["New Task"]
-      }
+        function: `${MODULE_ADDRESS}::todolist::delete_task`,
+        functionArguments: [taskToDelete?.task_id],
+      },
     });
 
     // using sign and submit separately
