+++ move/ts_scripts/src/fetch_data_from_chain.ts
@@ -1,4 +1,4 @@
-// npx ts-node ./ts/fetch_data_from_chain.ts
+// npx ts-node ./src/fetch_data_from_chain.ts
 import { type Task } from "./types";
 import { getTaskTableHandle } from "./get_task_table";
